+++ web/xfa_layer_builder.js
@@ -17,7 +17,7 @@
 // eslint-disable-next-line max-len
 /** @typedef {import("../src/display/annotation_storage").AnnotationStorage} AnnotationStorage */
 // eslint-disable-next-line max-len
-/** @typedef {import("../src/display/display_utils").PageViewport} PageViewport */
+/** @typedef {import("../src/display/page_viewport").PageViewport} PageViewport */
 /** @typedef {import("./pdf_link_service.js").PDFLinkService} PDFLinkService */
 
 import { XfaLayer } from "pdfjs-lib";
